modularizing pinecone and openai embeddings
